Use font awesome icons instead of svg assets
The previous svg assets were not accessible because high constrast
mode in Windows will omit background images. (I don't have Windows
and can't test this, but I believe the internets.) So our links
would be invisible in high contrast mode.

We could also have kept the previous images as content in a :before
pseudo element, however that was turing out to be a pain, and we've
switched to font awesome everywhere else, so I figured this woudl be
okay.

This commit also changes the alt text to aria labels. I didn't think
it was necessary to show those on hover. Hopefully the icons and href
are enough for people without screen readers.