Some font increases to make text easier to read on the default zoom level
Not strictly part of the accessibility audit, but I think the current
text is weirdly small/hard to read.
I don't plan on fixing the button contrast issue in this commit, because
I don't want to totally trash everyone's colour theme, so hopefully this
will make up for it a tiny bit.