Fix left & right padding on vertical forms
Adding padding to .form-vertical elements fixes the lack of padding
on the new restroom page and the contact page, where it needs to be
fixed, and doesn't add unwanted padding to other pages.

Doing this in _mobile.scss, because it only needs to apply at small
screen widths. Wider screen widths get padding one the sides of the
main.container element, so the extra padding isn't needed.