Add Vagrant catalog, update Vagrantfile
This should allow using our pre-provisioned Vagrant box.

(Downloads a box from Dropbox, as this is only a proof-of-concept.)

The catalog lives on GitHub.

Vagrant on your local commandline fetches the catalog, reads it,
and determines it needs to download the box from Dropbox.

The catalog also contains an SHA1 hash of the box to be downloaded,
so as to verify whether it has downloaded correctly.

This skips most of the provisioning steps, as the "idempotent" setup
script can tell that just about all the provisioning steps are already
completed.

Looking forward (regarding a maintenance strategy):...

As the project is updated, minor edits (especially to the "front-end")
will work fine with this pre-provisioned box... but major updates
to the project might require a new box to be produced.

Perhaps a new box should be made once every month, or every four
months?

(Also, we can easily move back to the non-pre-provisioned boxes,
either temporarily, or on a long-term basis).

Producing a new pre-provisioned box every once in a while will
keep provision times short, and allow download sizes from our box
hosting service to be as small as possible.