.rubocop.yml: Disable Style/CommentAnnotation rule
This rule was flagging the comment in app/models/restroom.rb
as being incorrect, due to it having "# Note:" on one line,
then the rest of the note/comments on the following lines.

Rather than reformat the first line of the comment to be
on the same line as the "# Note:", which IMO would not be as readable,
I have instead disabled the style rule in the Rubocop configuration.